Cleanup: List Helper Utilities
Summary:
	- Consolidated toList and toArray methods and calls into utility
	  class.
