+++ js/algo-challenges/maximum-twin-sum-of-a-linked-list/__tests__/maximum-twin-sum-of-a-linked-list.js
@@ -1,57 +1,21 @@
 const {pairSum} = require('../index');
+const {toList} = require('../../../test-utilities/linked-lists');
 
 describe('Sentence Similarity', function () {
-  /**
-   * @param {number[]} arr
-   * @return {ListNode}
-   */
-  function toList(arr) {
-    const head = {next: null};
-
-    let curr = head;
-    for (const val of arr) {
-      const newNode = {val, next: null};
-      curr.next = newNode;
-      curr = newNode;
-    }
-
-    return head.next;
-  }
-
-  function toArray(listHead) {
-    const arr = [];
-
-    let curr = listHead;
-    while (curr) {
-      arr.push(curr.val);
-      curr = curr.next;
-    }
-
-    return arr;
-  }
 
   it(`[5, 4, 2, 1] => 6`, function () {
-    let head = toList([5, 4, 2, 1]);
-    expect(toArray(head)).toEqual([5, 4, 2, 1]);
-    const result = pairSum(head);
+    const result = pairSum(toList([5, 4, 2, 1]));
     expect(result).toEqual(6);
-    // expect(toArray(head)).toEqual([5, 4, 2, 1]);
   });
 
   it(`[4, 2, 2, 3] => 7`, function () {
-    let head = toList([4, 2, 2, 3]);
-    expect(toArray(head)).toEqual([4, 2, 2, 3]);
-    const result = pairSum(head);
+    const result = pairSum(toList([4, 2, 2, 3]));
     expect(result).toEqual(7);
-    // expect(toArray(head)).toEqual([4, 2, 2, 3]);
   });
 
   it(`[1, 100000] => 100001`, function () {
-    let head = toList([1, 100000]);
-    expect(toArray(head)).toEqual([1, 100000]);
-    const result = pairSum(head);
+    const result = pairSum(toList([1, 100000]));
     expect(result).toEqual(100001);
-    // expect(toArray(head)).toEqual([1, 100000]);
   });
 
   it(`[1, 2, 3, 4, 5, 6] => 7`, function () {
