+++ js/algo-challenges/remove-duplicates-from-sorted-list-ii/__tests__/remove-duplicates-from-sorted-list-ii.tests.js
@@ -1,25 +1,5 @@
 const {deleteDuplicates} = require('../index');
-
-const toList = function (arr) {
-  if (!arr || arr.length === 0) {
-    return null;
-  }
-
-  const head = {val: arr[0], next: null};
-
-  let curr = head;
-  for (let i = 1; i < arr.length; i++) {
-    const newNode = {
-      val: arr[i],
-      next: null
-    };
-
-    curr.next = newNode;
-    curr = curr.next;
-  }
-
-  return head;
-};
+const {toList} = require('../../../test-utilities/linked-lists');
 
 describe('Remove Duplicates from Sorted List II', function () {
   it(`[1, 2, 3, 3, 4, 4, 5] => [1, 2, 5]`, function () {
