+++ js/algo-challenges/remove-duplicates-from-sorted-list-ii/index.js
@@ -5,6 +5,7 @@
  *     this.next = (next===undefined ? null : next)
  * }
  */
+
 /**
  * @param {ListNode} head
  * @return {ListNode}
