+++ js/algo-challenges/remove-nth-node-from-end-of-list/index.js
@@ -5,6 +5,7 @@
  *     this.next = (next===undefined ? null : next)
  * }
  */
+
 /**
  * @param {ListNode} head
  * @param {number} n
