+++ js/algo-challenges/reverse-linked-list-ii/__tests__/reverse-linked-list-ii.tests.js
@@ -1,35 +1,7 @@
 const {reverseBetween} = require('../index');
+const {toList} = require('../../../test-utilities/linked-lists');
 
 describe('Reverse Linked List II', function () {
-  /**
-   * @param {number[]} arr
-   * @return {ListNode}
-   */
-  function toList(arr) {
-    const head = {next: null};
-
-    let curr = head;
-    for (const val of arr) {
-      const newNode = {val, next: null};
-      curr.next = newNode;
-      curr = newNode;
-    }
-
-    return head.next;
-  }
-
-  function toArray(listHead) {
-    const arr = [];
-
-    let curr = listHead;
-    while (curr) {
-      arr.push(curr.val);
-      curr = curr.next;
-    }
-
-    return arr;
-  }
-
   it(`([1, 2, 3, 4, 5]), 2, 4) => [1, 4, 3, 2, 5]`, function () {
     const result = reverseBetween(toList([1, 2, 3, 4, 5]), 2, 4);
     expect(result).toEqual(toList([1, 4, 3, 2, 5]));
