+++ js/algo-challenges/reverse-linked-list/__tests__/reverse-linked-list.tests.js
@@ -1,26 +1,7 @@
 const {reverseList} = require('../index');
+const {toList} = require('../../../test-utilities/linked-lists');
 
 describe('Reverse Linked List', function () {
-  /**
-   * @param {number[]} arr
-   * @return {ListNode}
-   */
-  function toList(arr) {
-    const head = {next: null};
-
-    let curr = head;
-    for (const item of arr) {
-      const newNode = {
-        val: item,
-        next: null
-      };
-      curr.next = newNode;
-      curr = curr.next;
-    }
-
-    return head.next;
-  }
-
   it(`[1, 2, 3, 4, 5] => [5, 4, 3, 2, 1]`, function () {
     const result = reverseList(toList([1, 2, 3, 4, 5]));
     expect(result).toEqual(toList([5, 4, 3, 2, 1]));
